chore(zones): Nest ZoneIngresses and ZoneEgresses into their Zone
