+++ src/app/zone-egresses/views/item/ClustersView.vue
