+++ src/app/zone-egresses/views/item/StatsView.vue
