+++ src/app/zone-egresses/views/item/XdsConfigView.vue
