+++ src/app/zone-ingresses/views/item/StatsView.vue
