+++ src/app/zones/sources.ts
@@ -1,9 +1,10 @@
 import { DataSourceResponse } from '@/app/application/services/data-source/DataSourcePool'
+import { sources as zoneEgresses } from '@/app/zone-egresses/sources'
+import { sources as zoneIngresses } from '@/app/zone-ingresses/sources'
 import type KumaApi from '@/services/kuma-api/KumaApi'
 import type { PaginatedApiListResponse as CollectionResponse } from '@/types/api.d'
-import type { ZoneOverview, ZoneIngressOverview, ZoneEgressOverview, ZoneIngress, ZoneEgress } from '@/types/index.d'
+import type { ZoneOverview } from '@/types/index.d'
 export type { ZoneOverview } from '@/types/index.d'
-
 type PaginationParams = {
   size: number
   page: number
@@ -13,30 +14,19 @@ type DetailParams = {
   name: string
 }
 
-type EnvoyDataParams = DetailParams & {
-  dataPath: 'xds' | 'clusters' | 'stats'
-}
-
 type Closeable = { close: () => void }
 
 export type ZoneOverviewCollection = CollectionResponse<ZoneOverview>
 export type ZoneOverviewSource = DataSourceResponse<ZoneOverview>
 export type ZoneOverviewCollectionSource = DataSourceResponse<ZoneOverviewCollection>
 
-export type ZoneIngressSource = DataSourceResponse<ZoneIngress>
-export type ZoneIngressOverviewCollection = CollectionResponse<ZoneIngressOverview>
-export type ZoneIngressOverviewSource = DataSourceResponse<ZoneIngressOverview>
-export type ZoneIngressOverviewCollectionSource = DataSourceResponse<ZoneIngressOverviewCollection>
-
-export type ZoneEgressSource = DataSourceResponse<ZoneEgress>
-export type ZoneEgressOverviewCollection = CollectionResponse<ZoneEgressOverview>
-export type ZoneEgressOverviewSource = DataSourceResponse<ZoneEgressOverview>
-export type ZoneEgressOverviewCollectionSource = DataSourceResponse<ZoneEgressOverviewCollection>
-
 export type EnvoyDataSource = DataSourceResponse<object | string>
 
 export const sources = (api: KumaApi) => {
   return {
+    ...zoneIngresses(api),
+    ...zoneEgresses(api),
+
     '/zone-cps': async (params: PaginationParams, source: Closeable) => {
       source.close()
 
@@ -54,70 +44,5 @@ export const sources = (api: KumaApi) => {
       return await api.getZoneOverview({ name })
     },
 
-    '/zone-ingresses/:name': async (params: DetailParams, source: Closeable) => {
-      source.close()
-
-      const { name } = params
-
-      return await api.getZoneIngress({ name })
-    },
-
-    '/zone-ingresses/:name/data-path/:dataPath': (params: EnvoyDataParams, source: Closeable) => {
-      source.close()
-
-      const { name, dataPath } = params
-
-      return api.getZoneIngressData({ zoneIngressName: name, dataPath })
-    },
-
-    '/zone-ingress-overviews': async (params: PaginationParams, source: Closeable) => {
-      source.close()
-
-      const { size } = params
-      const offset = params.size * (params.page - 1)
-
-      return await api.getAllZoneIngressOverviews({ size, offset })
-    },
-
-    '/zone-ingress-overviews/:name': async (params: DetailParams, source: Closeable) => {
-      source.close()
-
-      const { name } = params
-
-      return await api.getZoneIngressOverview({ name })
-    },
-
-    '/zone-egresses/:name': async (params: DetailParams, source: Closeable) => {
-      source.close()
-
-      const { name } = params
-
-      return await api.getZoneEgress({ name })
-    },
-
-    '/zone-egresses/:name/data-path/:dataPath': (params: EnvoyDataParams, source: Closeable) => {
-      source.close()
-
-      const { name, dataPath } = params
-
-      return api.getZoneEgressData({ zoneEgressName: name, dataPath })
-    },
-
-    '/zone-egress-overviews': async (params: PaginationParams, source: Closeable) => {
-      source.close()
-
-      const { size } = params
-      const offset = params.size * (params.page - 1)
-
-      return await api.getAllZoneEgressOverviews({ size, offset })
-    },
-
-    '/zone-egress-overviews/:name': async (params: DetailParams, source: Closeable) => {
-      source.close()
-
-      const { name } = params
-
-      return await api.getZoneEgressOverview({ name })
-    },
   }
 }
