+++ src/app/zones/views/item/ConfigView.vue
