+++ src/app/zones/views/item/DetailView.vue
@@ -93,7 +93,7 @@ import {
   getZoneControlPlaneStatus,
   getZoneDpServerAuthType,
   getZoneControlPlaneEnvironment,
-} from '../data'
+} from '../../data'
 import DefinitionCard from '@/app/common/DefinitionCard.vue'
 import StatusBadge from '@/app/common/StatusBadge.vue'
 import SubscriptionList from '@/app/common/subscriptions/SubscriptionList.vue'
