+++ src/test-support/fake.ts
@@ -49,6 +49,7 @@ export type MockEnvKeys = keyof {
   KUMA_ZONEEGRESS_COUNT: string
   KUMA_ZONEINGRESS_COUNT: string
   KUMA_ZONE_COUNT: string
+  KUMA_ZONE_NAME: string
   KUMA_MESH_COUNT: string
   KUMA_SUBSCRIPTION_COUNT: string
   KUMA_GLOBALSECRET_COUNT: string
